8134540: Much nearly duplicated code for PerfMemory support
