+++ src/hotspot/os/aix/os_aix.inline.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1999, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1999, 2021, Oracle and/or its affiliates. All rights reserved.
  * Copyright (c) 2012, 2018 SAP SE. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
@@ -69,18 +69,6 @@ inline int os::ftruncate(int fd, jlong length) {
   return ::ftruncate64(fd, length);
 }
 
-// macros for restartable system calls
-
-#define RESTARTABLE(_cmd, _result) do { \
-    _result = _cmd; \
-  } while(((int)_result == OS_ERR) && (errno == EINTR))
-
-#define RESTARTABLE_RETURN_INT(_cmd) do { \
-  int _result; \
-  RESTARTABLE(_cmd, _result); \
-  return _result; \
-} while(false)
-
 // We don't have NUMA support on Aix, but we need this for compilation.
 inline bool os::numa_has_static_binding()   { ShouldNotReachHere(); return true; }
 inline bool os::numa_has_group_homing()     { ShouldNotReachHere(); return false;  }
@@ -89,10 +77,6 @@ inline size_t os::write(int fd, const void *buf, unsigned int nBytes) {
   size_t res;
   RESTARTABLE((size_t) ::write(fd, buf, (size_t) nBytes), res);
   return res;
-}
-
-inline int os::close(int fd) {
-  return ::close(fd);
 }
 
 inline int os::socket_close(int fd) {
